requirements.txt
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
python-dotenv==1.0.1
azure-identity==1.16.0
gunicorn==22.0.0
fastapi==0.115.4
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-multipart==0.0.17
qdrant-client==1.12.1
langfuse==2.53.9
pypdf==5.1.0
unstructured==0.15.13